[new-server-side-architecture] - Remove code duplication
