+++ akka-bbb-apps/src/main/scala/org/bigbluebutton/core/apps/plugin/PluginHdlrHelpers.scala
@@ -2,7 +2,8 @@ package org.bigbluebutton.core.apps.plugin
 
 import org.bigbluebutton.common2.msgs.PluginDataChannelReplaceOrDeleteBaseBody
 import org.bigbluebutton.core.db.PluginDataChannelEntryDAO
-import org.bigbluebutton.core.models.{ Roles, UserState }
+import org.bigbluebutton.core.models.{ DataChannel, PluginModel, Roles, UserState, Users2x }
+import org.bigbluebutton.core.running.LiveMeeting
 
 object PluginHdlrHelpers {
   def checkPermission(user: UserState, permissionType: List[String], creatorCheck: => Boolean = false): List[Boolean] = {
@@ -23,5 +24,27 @@ object PluginHdlrHelpers {
       msgBody.entryId
     )
     creatorUserId == userId
+  }
+
+  def dataChannelCheckingLogic(liveMeeting: LiveMeeting, userId: String,
+                               pluginName: String, channelName: String,
+                               caseSomeDataChannelAndPlugin: (UserState, DataChannel, String) => Unit): Option[Unit] = {
+    val pluginsDisabled: Boolean = liveMeeting.props.meetingProp.disabledFeatures.contains("plugins")
+    val meetingId = liveMeeting.props.meetingProp.intId
+
+    for {
+      _ <- if (!pluginsDisabled) Some(()) else None
+      user <- Users2x.findWithIntId(liveMeeting.users2x, userId)
+    } yield {
+      PluginModel.getPluginByName(liveMeeting.plugins, pluginName) match {
+        case Some(p) =>
+          p.manifest.content.dataChannels.getOrElse(List()).find(dc => dc.name == channelName) match {
+            case Some(dc) =>
+              caseSomeDataChannelAndPlugin(user, dc, meetingId)
+            case None => println(s"Data channel '${channelName}' not found in plugin '${pluginName}'.")
+          }
+        case None => println(s"Plugin '${pluginName}' not found.")
+      }
+    }
   }
 }
